fix: ErrNoObjects from GetDescriptor(s)
When GetDescriptor/GetDescriptors is called on an image with no
descriptors, the returned error can be confusing. For example, when a
DescriptorSelectorFunc is used with an invalid argument against an image
with no descriptors, ErrObjectNotFound would be returned. ErrNoObjects
allows the caller to differentiate this case, and display a more
descriptive error.