refactor: improve compaction logic for AddObject
Previously, an add operation would simply append the data object to the
end of the data section. If one or more object(s) had previously been
deleted from the image, this could result in wasted space. Now, we trim
the data section based on in-use objects before adding the new object.
Update tests to reflect this new behaviour.